+++ lib/constants/environment_config.dart
@@ -1,0 +1,5 @@
+
+class EnvironmentConfig {
+  static const nightlyBuildVersion =
+  String.fromEnvironment('nightlyBuildVersion', defaultValue: "");
+}
