--- lib/providers/environment_provider.dart
@@ -1,8 +1,0 @@
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-final environmentProvider = Provider((ref) => EnvironmentConfig());
-
-class EnvironmentConfig {
-  static const nightlyBuildVersion =
-      String.fromEnvironment('nightlyBuildVersion', defaultValue: "");
-}
